rename classes b27
