+++ src/components/abstract-component.js
