+++ src/components/films-sub-section.js
@@ -1,4 +1,4 @@
-import AbstractComponent from "./abstract-class";
+import AbstractComponent from "./abstract-component";
 
 const createFilmExtraListTemplate = (title, className) => {
   return (
