+++ src/components/sort-form.js
@@ -1,4 +1,4 @@
-import AbstractComponent from "./abstract-class";
+import AbstractComponent from "./abstract-component";
 
 const createSortFormTemplate = () => {
   return (
