+++ src/main.js
@@ -5,7 +5,7 @@ import SortFormComponent from './components/sort-form.js';
 import FilmSectionComponent from './components/films-section.js';
 import {generateFilmCards} from './mocks/film-card.js';
 import {RenderPosition, render} from './utils.js/render.js';
-import PageController from './controllers/film-section.js';
+import PageController from './controllers/page-controller';
 
 const TOTAL_AMOUNT_CARDS = 15;
 const WATCHED_MOVIES = 15;
